Got importing TeX to work in Prince2. The home for the pretex.exe program and the mmldll.dll and .gmr files is xpi-stage\prince, and the ptdata directory is directly below that.
The build process for filters is better, but still broken. I couldn't get funnelweb to accept its "include" directive, which requires all the .fw files to be copied to obj-prince\filters\ptprince. All of obj-prince\filters\ptprince needs to be cleaned out (except for makefile) to rebuild from the funnelweb sources.