I think I have finally fixed the problems we've been having with finding properties and methods on VCam objects. The problem has been that in the JavaScript code, the expression (for example) obj.readyState will return 0, 1, or 2, but more frequently it returns void. What really made it weird was that even when it returned void, I could evaluate it in the debugger, and it would succeed. This meant the object was there and the correct object; it just wouldn't evaluate properties or methods. It would return correct values for its id, tag name, etc. at all times.
The crucial clue (after many dead ends) was that I noticed that sometimes in the debugger (Venkman) obj.toString() would return ""[object XPCNativeWrapper [object HTMLObjectElement ..]]" and at other times it would return "[[HTMLObjectElement] [class: HTMLObjectElement...]]". In the latter case, everything worked correctly, and in the former it always failed, so it was time to research what XPCNativeWrapper *really* does.

It is there partly for security, to keep an evil web server from modifying methods on objects in the content document that would be executed in the chrome (main program) code. So a wrapped HTML object element supports only those methods that are in the IDL file for HTMLObjectElement, and it goes to the code in the C++ implementation, skipping over anything defined in JavaScript. (see http://https://developer.mozilla.org/en-US/docs/Xray_vision for more details). It also skips over any properties and methods defined in the plugin.

This is why some methods, such as obj.id, obj.tagName, etc. worked while all the properties and methods built into the VCam object failed when obj was wrapped.

There is a way around this. Given a wrapped object, the .wrappedJSObject property is supposed to return the underlying object, unwrapped. Unfortunately, when you evaluate obj.wrappedJSObject in the debugger, you get void (another curve ball). If you ignore that and execute

if (obj.wrappedJSObject) obj=obj.wrappedJSObject;

in the program, the condition evaluates to true, despite what the debugger said. The above line of code doesn't return an HTMLObjectElement, but a [XPCSafeJSObjectWrapper], but it works. So to fix the code, all that was necessary was the above line in a few key places. In retrospect, the only times that the VCam methods worked was just after the HTML object element was created and was still being held in memory. They failed every time the object was constructed using the object node in the document (which is always when you have loaded a document with plots). The debugger evidently finds the wrapped object when it evaluates something, which is why everything always worked in the debugger.

We can now print plots. There are some gross scaling and placement problems, but they are easily fixable. I can breathe easier now.