On the Mac, there has been a problem with entering macros using ctrl+space. It should be cmd+space, but that keyboard shortcut defaults to spotlight, the Mac's system search. Many utilities like Alfred and Launchbar will use ctrl+space. Up until now we have used ctrl+space, but that is frequently already taken, so the macro feature is unavailable.
I have now made it so the follow keyboard shortcuts will work if they are available:

	ctrl+space
	ctrl+alt+space
	ctrl+alt+cmd+space, and
	shift+ctrl+alt+cmd+space

All the modifier keys are near each other, so these combinations are actually accessible. So now the story is:

On the Mac, try ctrl+space. If that doesn't work, go down this list until you find something that does.

I also made it so the escape key gets you out.